added program option '--exclude-file'
