+++ examples/CMakeLists.txt
@@ -3,6 +3,7 @@ set(target example)
 
 add_executable(${target} 
   ${CMAKE_CURRENT_SOURCE_DIR}/step_definition.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/step_custom_parameters.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/hooks.cpp
 )
 install(
